move class into `PHPStan\Php` namespace
